FEAT: Criando o arquivo de TypePID
