+++ sdkconfig.ESP32-BraiaV3
@@ -580,6 +580,13 @@ CONFIG_ESP_PHY_REDUCE_TX_POWER=y
 # end of Power Management
 
 #
+# ESP Ringbuf
+#
+# CONFIG_RINGBUF_PLACE_FUNCTIONS_INTO_FLASH is not set
+# CONFIG_RINGBUF_PLACE_ISR_FUNCTIONS_INTO_FLASH is not set
+# end of ESP Ringbuf
+
+#
 # ESP System Settings
 #
 # CONFIG_ESP_SYSTEM_PANIC_PRINT_HALT is not set
@@ -662,6 +669,7 @@ CONFIG_ESP32_WIFI_ENABLE_WPA3_SAE=y
 CONFIG_ESP_WIFI_STA_DISCONNECTED_PM_ENABLE=y
 # CONFIG_ESP_WIFI_GMAC_SUPPORT is not set
 CONFIG_ESP_WIFI_SOFTAP_SUPPORT=y
+# CONFIG_ESP_WIFI_SLP_BEACON_LOST_OPT is not set
 # end of Wi-Fi
 
 #
@@ -1321,6 +1329,7 @@ CONFIG_WIFI_PROV_AUTOSTOP_TIMEOUT=30
 # CONFIG_WIFI_PROV_BLE_BONDING is not set
 CONFIG_WIFI_PROV_BLE_SEC_CONN=y
 CONFIG_WIFI_PROV_BLE_FORCE_ENCRYPTION=y
+# CONFIG_WIFI_PROV_KEEP_BLE_ON_AFTER_PROV is not set
 # end of Wi-Fi Provisioning Manager
 
 #
